+++ MineCloud-Service-Info.ts
@@ -1,9 +1,11 @@
+import 'dotenv/config'
+
 // AWS
-export const AWS_ACCOUNT_ID = '';
-export const AWS_REGION = '';
+export const AWS_ACCOUNT_ID = process.env.AWS_ACCOUNT_ID || '';
+export const AWS_REGION = process.env.AWS_REGION || '';
 
 // Discord
-export const DISCORD_APP_ID = '';
-export const DISCORD_PUBLIC_KEY = '';
-export const DISCORD_BOT_TOKEN = '';
-export const DISCORD_CHANNEL_WEB_HOOK = '';
+export const DISCORD_APP_ID = process.env.DISCORD_APP_ID || '';
+export const DISCORD_PUBLIC_KEY = process.env.DISCORD_PUBLIC_KEY || '';
+export const DISCORD_BOT_TOKEN = process.env.DISCORD_BOT_TOKEN || '';
+export const DISCORD_CHANNEL_WEB_HOOK = process.env.DISCORD_CHANNEL_WEB_HOOK || '';