+++ minecloud_configuration_packages/ArkSurvivalEvolved/dev/default/ark-server-settings.ts
@@ -1,0 +1,3 @@
+export const ARK_SERVER_NAME = 'MC_ARK';
+export const ARK_WORLD = 'MC_DEDICATED';
+export const ARK_PASSWORD = 'secret';